move all Title logic into delta_patterns.py
